+++ javascript/libjoynr-js/src/main/js/joynr/dispatching/types/SubscriptionPublication.js
@@ -1,3 +1,5 @@
+/*jslint node: true */
+
 /*
  * #%L
  * %%
@@ -16,11 +18,9 @@
  * limitations under the License.
  * #L%
  */
-
-define("joynr/dispatching/types/SubscriptionPublication", [
-    "joynr/util/UtilInternal",
-    "joynr/util/Typing"
-], function(Util, Typing) {
+var UtilInternal = require('../../util/UtilInternal');
+var Typing = require('../../util/Typing');
+module.exports = (function(Util, Typing) {
 
     /**
      * @name SubscriptionPublication
@@ -67,4 +67,4 @@ define("joynr/dispatching/types/SubscriptionPublication", [
 
     return SubscriptionPublication;
 
-});
+}(UtilInternal, Typing));