+++ javascript/libjoynr-js/src/main/js/joynr/messaging/routing/MessageQueue.js
@@ -1,3 +1,5 @@
+/*jslint node: true */
+
 /*
  * #%L
  * %%
@@ -16,21 +18,17 @@
  * limitations under the License.
  * #L%
  */
-
 /**
  * The <code>MessageQueue</code> is a joynr internal data structure. The Message Queue caches incoming messages, which cannot be shipped
  * to the correct participant. Once a participant with the matching participantId is registered, the incoming message is forwarded to him
  */
-define(
-        "joynr/messaging/routing/MessageQueue",
-        [
-            "joynr/system/LoggerFactory",
-            "joynr/system/DiagnosticTags",
-            "joynr/util/LongTimer",
-            "joynr/util/UtilInternal",
-            "joynr/messaging/JoynrMessage"
-        ],
-        function(LoggerFactory, DiagnosticTags, LongTimer, Util, JoynrMessage) {
+var LoggerFactory = require('../../system/LoggerFactory');
+var DiagnosticTags = require('../../system/DiagnosticTags');
+var LongTimer = require('../../util/LongTimer');
+var UtilInternal = require('../../util/UtilInternal');
+var JoynrMessage = require('../JoynrMessage');
+module.exports =
+        (function(LoggerFactory, DiagnosticTags, LongTimer, Util, JoynrMessage) {
 
             var defaultSettings;
             /**
@@ -158,4 +156,4 @@ define(
             };
 
             return MessageQueue;
-        });
+        }(LoggerFactory, DiagnosticTags, LongTimer, UtilInternal, JoynrMessage));