+++ javascript/libjoynr-js/src/main/js/joynr/proxy/SubscriptionQos.js
@@ -1,3 +1,5 @@
+/*jslint node: true */
+
 /*
  * #%L
  * %%
@@ -16,15 +18,11 @@
  * limitations under the License.
  * #L%
  */
-
-define(
-        "joynr/proxy/SubscriptionQos",
-        [
-            "joynr/util/Typing",
-            "joynr/util/UtilInternal",
-            "joynr/system/LoggerFactory"
-        ],
-        function(Typing, Util, LoggerFactory) {
+var Typing = require('../util/Typing');
+var UtilInternal = require('../util/UtilInternal');
+var LoggerFactory = require('../system/LoggerFactory');
+module.exports =
+        (function(Typing, Util, LoggerFactory) {
 
             var defaultSettings;
 
@@ -237,4 +235,4 @@ define(
 
             return SubscriptionQos;
 
-        });
+        }(Typing, UtilInternal, LoggerFactory));